+++ airflow/example_dags/example_asset_decorator.py
@@ -18,9 +18,9 @@ from __future__ import annotations
 
 import pendulum
 
-from airflow.assets import Asset
 from airflow.decorators import dag, task
-from airflow.decorators.assets import asset
+from airflow.sdk.definitions.asset import Asset
+from airflow.sdk.definitions.decorators import asset
 
 
 @asset(uri="s3://bucket/asset1_producer", schedule=None)
